フックを作成・呼び出し
