+++ src/App.tsx
@@ -1,38 +1,11 @@
 import "./styles.css";
 import { UserCard } from "./components/UserCard";
-import axios from "axios";
-import { User } from "./types/api/user";
-import { UserProfile } from "./types/UserProfile";
-import { useState } from "react";
+import { useAllUsers } from "./hooks/useAllUsers";
 
 export default function App() {
-  const [userProfiles, setUserProfiles] = useState<Array<UserProfile>>([]);
+  const { getUsers, userProfiles, loading, error } = useAllUsers();
 
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(false);
-
-  const OnClickFetchUser = () => {
-    setLoading(true);
-    setError(false);
-    axios
-      .get<Array<User>>("https://jsonplaceholder.typicode.com/users")
-      .then((res) => {
-        const data = res.data.map((user) => ({
-          id: user.id,
-          name: `${user.name}(${user.username})`,
-          email: user.email,
-          address: `${user.address.city}${user.address.suite}${user.address.street}`
-        }));
-
-        setUserProfiles(data);
-      })
-      .catch(() => {
-        setError(true);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
+  const OnClickFetchUser = () => getUsers();
   return (
     <div className="App">
       <button onClick={OnClickFetchUser}>データー取得</button>
